(feat) merge tab layout and common layouts
