+++ src/routes/(tab)/+page.svelte
@@ -5,10 +5,10 @@
 </script>
 
 <svelte:head>
-    <meta property="og:title" content="홈 | 종이상자 공간">
-    <meta property="og:description" content="진행 중인 프로젝트, 활동 및 소식을 확인하세요">
-    <meta name="description" content="진행 중인 프로젝트, 활동 및 소식을 확인하세요">
-    <title>홈 | 종이상자 공간</title>
+    <meta property="og:title" content="종이상자 공간">
+    <meta property="og:description" content="진행 중인 프로젝트, 활동 및 소식">
+    <meta name="description" content="진행 중인 프로젝트, 활동 및 소식">
+    <title>종이상자 공간</title>
 </svelte:head>
 
 <main>
